requirements.txt
-e git+https://github.com/slugnasty/bepto-engine.git@a38b0a9d69fc1075a62ec4aa4f35cb5aabf4e880#egg=bengine&subdirectory=bengine
cx-Freeze==6.10
cx-Logging==3.0
glfw==2.5.0
multipledispatch==0.6.0
numpy==1.22.2
Pillow==9.0.1
PyOpenGL==3.1.5
pyrr==0.10.3
six==1.16.0
